% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_book_details}
\alias{get_book_details}
\title{get_book_details on page}
\usage{
get_book_details(library_html, book_html)
}
\arguments{
\item{library_html}{the result of rvest::read_html() of class xml_document xml_node}

\item{book_html}{each book html from lapply(X = get_links(page_html), FUN = rvest::read_html)}
}
\value{
dataframe
}
\description{
get_book_details on page
}
\examples{

res <- get_book_details(library_html = library_html, book_html = book_html)

}
