% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_ISBN}
\alias{get_ISBN}
\title{get_ISBN}
\usage{
get_ISBN(book_html)
}
\arguments{
\item{book_html}{the result of rvest::read_html() of class xml_document xml_node}
}
\value{
character, ISBN number
}
\description{
get_ISBN
}
\examples{
isbn <- get_ISBN(url = url)

}
