% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_last_page}
\alias{get_last_page}
\title{get_last_page}
\usage{
get_last_page(html)
}
\arguments{
\item{html}{the result of rvest::read_html() of class xml_document xml_node}
}
\value{
numeric value of last page
}
\description{
get_last_page
}
\examples{

URL <- "https://lubimyczytac.pl/ksiegozbior/xc5GvparjG"
html <- rvest::read_html(x = URL)
last_page <- get_last_page(html = html)

}
