% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run_libroScrapeR}
\alias{run_libroScrapeR}
\title{run_libroScrapeR}
\usage{
run_libroScrapeR(URL)
}
\arguments{
\item{URL}{}
}
\value{
dataframe
}
\description{
run_libroScrapeR
}
\examples{

res <- run_libroScrapeR(URL = URL)

}
